+++ src/Vault.sol
@@ -30,7 +30,7 @@ contract Vault {
 
     struct YStake {
         address user;
-        uint192 strike;
+        uint128 strike;
         uint32 epochId;
         uint256 amount;
         uint256 claimed;
@@ -43,7 +43,7 @@ contract Vault {
 
     struct HodlStake {
         address user;
-        uint192 strike;
+        uint128 strike;
         uint32 epochId;
         uint256 amount;
     }
@@ -56,7 +56,7 @@ contract Vault {
     uint256 public yieldPerTokenAcc;
     uint256 public cumulativeYieldAcc;
     struct EpochInfo {
-        uint192 strike;
+        uint128 strike;
         uint256 yieldPerTokenAcc;
         uint256 cumulativeYieldAcc;
     }
@@ -65,9 +65,8 @@ contract Vault {
     // Map strike to active epoch ID
     mapping (uint256 => uint32) public epochs;
 
-
     // Events
-    event Triggered(uint192 indexed strike,
+    event Triggered(uint128 indexed strike,
                     uint32 indexed epoch,
                     uint256 timestamp);
 
@@ -76,27 +75,27 @@ contract Vault {
                uint256 amount);
 
     event HodlStaked(address indexed user,
-                     uint192 indexed strike,
+                     uint128 indexed strike,
                      uint32 indexed stakeId,
                      uint256 amount);
 
     event HodlUnstaked(address indexed user,
-                       uint192 indexed strike,
+                       uint128 indexed strike,
                        uint32 indexed stakeId,
                        uint256 amount);
 
     event HodlRedeemed(address indexed user,
-                       uint192 indexed strike,
+                       uint128 indexed strike,
                        uint32 indexed stakeId,
                        uint256 amount);
 
     event YStaked(address indexed user,
-                  uint192 indexed strike,
+                  uint128 indexed strike,
                   uint32 indexed stakeId,
                   uint256 amount);
 
     event YUnstaked(address indexed user,
-                    uint192 indexed strike,
+                    uint128 indexed strike,
                     uint32 indexed stakeId,
                     uint256 amount);
 
@@ -111,7 +110,7 @@ contract Vault {
         yMulti = new YMultiToken("", address(this));
     }
 
-    function deployERC20(uint192 strike) public returns (address) {
+    function deployERC20(uint128 strike) public returns (address) {
         if (deployments[strike] != address(0)) {
             return deployments[strike];
         }
@@ -139,7 +138,7 @@ contract Vault {
         cumulativeYieldAcc = total;
     }
 
-    function mint(uint192 strike) external payable {
+    function mint(uint128 strike) external payable {
         require(oracle.price(0) <= strike, "strike too low");
 
         uint256 before = stEth.balanceOf(address(this));
@@ -184,7 +183,7 @@ contract Vault {
         return false;
     }
 
-    function redeem(uint192 strike, uint256 amount, uint32 stakeId) external {
+    function redeem(uint128 strike, uint256 amount, uint32 stakeId) external {
         if (stakeId == 0) {
             // Redeem via tokens
             require(hodlMulti.balanceOf(msg.sender, strike) >= amount);
@@ -233,7 +232,7 @@ contract Vault {
         emit HodlRedeemed(msg.sender, strike, stakeId, amount);
     }
 
-    function yStake(uint192 strike, uint256 amount, address user) public returns (uint32) {
+    function yStake(uint128 strike, uint256 amount, address user) public returns (uint32) {
 
         require(yMulti.balanceOf(msg.sender, strike) >= amount, "y stake balance");
         uint32 epochId = epochs[strike];
@@ -316,7 +315,7 @@ contract Vault {
         claimed += amount;
     }
 
-    function hodlStake(uint192 strike, uint256 amount, address user) public returns (uint32) {
+    function hodlStake(uint128 strike, uint256 amount, address user) public returns (uint32) {
         require(hodlMulti.balanceOf(msg.sender, strike) >= amount, "hodl stake balance");
 
         hodlMulti.burn(msg.sender, strike, amount);
@@ -357,7 +356,7 @@ contract Vault {
         require(epochId < nextId, "invalid epoch");
 
         uint256 ypt;
-        uint192 strike = infos[epochId].strike;
+        uint128 strike = infos[epochId].strike;
         if (epochs[strike] == epochId) {
             // active epoch
             ypt = yieldPerToken() - infos[epochId].yieldPerTokenAcc;
